Updated cli
